refactor: deduplicate formatBytes, filterTree, downloadBlob across file browsers
- Move formatBytes to utils/format.ts as single source of truth
- Extract filterTree and downloadBlob into shared file-browser-utils.ts
- Remove 6 duplicate formatBytes definitions across renderers and modals
- ImageFilesystemModal now uses shared downloadBlob instead of inline logic
