+++ web/src/components/resources/renderers/HelmRepositoryRenderer.tsx
@@ -2,6 +2,7 @@ import { Database, AlertTriangle, CheckCircle2, Shield } from 'lucide-react'
 import { clsx } from 'clsx'
 import { Section, PropertyList, Property, ConditionsSection } from '../drawer-components'
 import { formatAge } from '../resource-utils'
+import { formatBytes } from '../../../utils/format'
 import { GitOpsStatusBadge, SyncCountdown } from '../../gitops'
 import { fluxConditionsToGitOpsStatus, type FluxCondition } from '../../../types/gitops'
 
@@ -163,12 +164,4 @@ export function HelmRepositoryRenderer({ data }: HelmRepositoryRendererProps) {
       <ConditionsSection conditions={conditions} />
     </>
   )
-}
-
-function formatBytes(bytes: number): string {
-  if (bytes === 0) return '0 B'
-  const k = 1024
-  const sizes = ['B', 'KB', 'MB', 'GB']
-  const i = Math.floor(Math.log(bytes) / Math.log(k))
-  return `${parseFloat((bytes / Math.pow(k, i)).toFixed(1))} ${sizes[i]}`
 }
