+++ web/src/components/resources/resource-utils.ts
@@ -725,13 +725,9 @@ export function formatDuration(ms: number, detailed: boolean = false): string {
   return '<1s'
 }
 
-export function formatBytes(bytes: number): string {
-  if (bytes === 0) return '0 B'
-  const k = 1024
-  const sizes = ['B', 'KB', 'MB', 'GB']
-  const i = Math.floor(Math.log(bytes) / Math.log(k))
-  return `${parseFloat((bytes / Math.pow(k, i)).toFixed(1))} ${sizes[i]}`
-}
+// Re-export from canonical location
+import { formatBytes as _formatBytes } from '../../utils/format'
+export const formatBytes = _formatBytes
 
 export function cronToHuman(cron: string): string {
   if (!cron) return '-'
